+++ ICD.Connect.Krang/KrangLicenseAttributes.cs
@@ -1,0 +1,8 @@
+﻿namespace ICD.Connect.Krang
+{
+	public static class KrangLicenseAttributes
+	{
+		public const string MAC_ADDRESS = "MacAddress";
+		public const string LICENSE_VERSION = "LicenseVersion";
+	}
+}