+++ CHANGELOG.md
@@ -6,6 +6,61 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [11.0.1] - 2018-11-20
+### Changed
+ - Fixed bug where USER migration would create redundant CommonConfig directory
+
+## [11.0.0] - 2018-11-08
+### Added
+ - Logging stacktraces when failing to dispose originators
+
+### Changed
+ - Potential performance improvement when removing settings
+
+### Removed
+ - Removed DestinationGroups
+
+## [10.0.0] - 2018-10-30
+### Added
+ - NVRAM is automatically migrated to USER directory
+
+## [9.0.0] - 2018-09-14
+### Changed
+ - Originators are constrained to class type
+ - Significant improvements to routing performance
+ - Only starting the direct message manager when broadcasting is configured
+
+## [8.1.0] - 2018-07-19
+### Added
+ - Added ActionSchedulerService to ServiceProvider
+
+## [8.0.2] - 2018-06-19
+### Changed
+ - Clear originators first, then dispose, fixing issue where id would be 0'ed out by dispose, causing the originators to never be cleared.
+ - Fixed bug with serialization of nested originator settings
+
+## [8.0.1] - 2018-06-04
+### Changed
+ - Fixes for license encoding
+
+## [8.0.0] - 2018-05-24
+### Changed
+ - Changed SPlusOnKrangLoadedEventRelay to SPlusKrangEventRelay
+ 
+### Added
+ - Added OnKrangCleared event to SPlusKrangEventRelay for completeness
+ 
+### Removed
+ - Removed the SPlusPanelShim, as it doesnt make sense
+ - Removed the SPlusSwitcherShim, relocated to ICD.Connect.Routing
+ 
+### Changed
+ - Drastically increasing thread count for pooling
+
+## [7.0.0] - 2018-05-18
+### Added
+ - Adding license validation, prevent loading Krang config if license is invalid
+
 ## [6.0.0] - 2018-05-09
 ### Changed
  - Renamed ambiguous s+interface to shim